Исправляет разметку для реализации строки с количеством показанным коментариев
